DSL query refactoring

+++ src/main/scala/org/scalawiki/dto/Page.scala
@@ -27,7 +27,7 @@ case class Page(
 }
 
 object Page {
-  type Id = Int
+  type Id = Long
 
   def full(
             id: Id,
